fenrir_vk_2.101-Kotlin
fenrir_vk_2.101-Kotlin

+++ app_filegallery/src/main/kotlin/dev/ragnarok/filegallery/model/tags/TagOwner.kt
@@ -15,7 +15,7 @@ class TagOwner : Parcelable {
     constructor(`in`: Parcel) {
         id = `in`.readInt()
         name = `in`.readString()
-        count = `in`.readByte().toInt()
+        count = `in`.readInt()
     }
 
     override fun describeContents(): Int {
